% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_download.R
\name{image_download}
\alias{image_download}
\title{Download Image Raster for ArcGIS Online}
\usage{
image_download(
  webserver,
  bbox = NULL,
  bbox_sr = NULL,
  size = "600,800",
  verbose = FALSE
)
}
\arguments{
\item{webserver}{list output from \code{service_info}}

\item{bbox}{\code{NULL} or named list formatted like: \code{list(xmin = -123,
ymin = -123, xmax = 123, ymax = 123)}. If \code{NULL}, will default to the
intial extent specified in the source file.}

\item{bbox_sr}{\code{NULL}, default; or 4-digit wkid as character. If
\code{NULL}, will default to the spatial reference in the source file.}

\item{size}{character, defaults to "600,800"; change to what you need.}

\item{verbose}{Logical, defaults to \code{FALSE}}
}
\value{
RasterBrick
}
\description{
Download Image Raster for ArcGIS Online
}
