% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_download.R
\name{layer_download}
\alias{layer_download}
\title{Download layer from ArcGIS online}
\usage{
layer_download(layer_info, outfields = NULL, verbose = FALSE, query = NULL)
}
\arguments{
\item{layer_info}{output from \code{layer_info}}

\item{outfields}{character or \code{NULL} (default). Quoted text with comma
sperated field names. The default value of \code{NULL} will make the
function return all the fields.}

\item{verbose}{logical, defaults to \code{FALSE}. Passed to
\code{crul::HttpClient$new()}.}

\item{query}{character or \code{NULL} (default). The default \code{NULL} will
return all records. Use a standard SQL style query to return a subset of
records. For example: \code{"OBJECTID > 1 AND OBJECTID <30"}.}
}
\value{
sf
}
\description{
Download layer from ArcGIS online
}
