% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_info.R
\name{layer_info}
\alias{layer_info}
\title{Return layer information}
\usage{
layer_info(mapservice, layer_id)
}
\arguments{
\item{mapservice}{object output from service_info}

\item{layer_id}{numeric, specifying the layer id.}
}
\value{
list
}
\description{
Return layer information
}
