% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{make_request}
\alias{make_request}
\title{Make a request}
\usage{
make_request(
  url,
  path = NULL,
  args = list(),
  parse_json = TRUE,
  img = FALSE,
  ...
)
}
\arguments{
\item{url}{character, the base url}

\item{path}{character, anything after the url}

\item{args}{list, named query list}

\item{parse_json}{logical, probably \code{TRUE}}

\item{img}{logical, defaults to \code{FALSE}. Needs to be \code{TRUE} if downloading raster.}

\item{...}{curl options}
}
\value{
If \code{parse_json = TRUE} returns a list, else returns unparsed text.
}
\description{
Right now this works with json files
}
\keyword{internal}
