% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{p_encode_m}
\alias{p_encode_m}
\title{Modified percent encode}
\usage{
p_encode_m(query)
}
\arguments{
\item{query}{character string you want to encode.}
}
\value{
character string
}
\description{
Modified percent encode
}
\keyword{internal}
