% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_info.R
\name{service_info}
\alias{service_info}
\title{Return Mapserver Info}
\usage{
service_info(mapserver_url)
}
\arguments{
\item{mapserver_url}{character}
}
\value{
list
}
\description{
Return Mapserver Info
}
