% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_info.R
\name{service_info}
\alias{service_info}
\title{Return Mapserver Info}
\usage{
service_info(mapserver_url, verbose = FALSE)
}
\arguments{
\item{mapserver_url}{character}

\item{verbose}{logical passed to \code{crul::HttpClient$new()}}
}
\value{
list
}
\description{
Return Mapserver Info
}
